update from years of neglect

+++ wp-content/plugins/w3-total-cache/Cdn_Plugin.php
@@ -17,11 +17,13 @@ class Cdn_Plugin {
 	 * Config
 	 */
 	private $_config = null;
+	private $_debug = false;
 
-	private $_replaced_urls = array();
+	private $_attachments_action = array();
 
 	function __construct() {
 		$this->_config = Dispatcher::config();
+		$this->_debug = $this->_config->get_boolean( 'cdn.debug' );
 	}
 
 	/**
@@ -29,38 +31,18 @@ class Cdn_Plugin {
 	 */
 	function run() {
 		$cdn_engine = $this->_config->get_string( 'cdn.engine' );
-		if ( Cdn_Util::is_engine_fsd( $cdn_engine ) ) {
-			$this->run_fsd();
-			return;
-		}
 
 		add_filter( 'cron_schedules', array(
 				$this,
 				'cron_schedules'
 			) );
 
-		if ( !$this->_config->get_boolean( 'cdn.debug' ) )
-			add_filter( 'w3tc_footer_comment', array(
-					$this,
-					'w3tc_footer_comment'
-				) );
+		add_filter( 'w3tc_footer_comment', array(
+				$this,
+				'w3tc_footer_comment'
+			) );
 
 		if ( !Cdn_Util::is_engine_mirror( $cdn_engine ) ) {
-			add_action( 'delete_attachment', array(
-					$this,
-					'delete_attachment'
-				) );
-
-			add_filter( 'update_attached_file', array(
-					$this,
-					'update_attached_file'
-				) );
-
-			add_filter( 'wp_update_attachment_metadata', array(
-					$this,
-					'update_attachment_metadata'
-				) );
-
 			add_action( 'w3_cdn_cron_queue_process', array(
 					$this,
 					'cron_queue_process'
@@ -80,7 +62,27 @@ class Cdn_Plugin {
 					$this,
 					'update_feedback'
 				) );
+
 		}
+
+		$flush_on_actions = !$this->_config->get_boolean( 'cdn.flush_manually' );
+
+		if ( $flush_on_actions ) {
+			add_action( 'delete_attachment',
+				array( $this, 'delete_attachment' ) );
+
+			add_filter( 'wp_insert_attachment_data',
+				array( $this, 'check_inserting_new_attachment' ), 10, 2 );
+
+			add_filter( 'update_attached_file',
+				array( $this, 'update_attached_file' ) );
+
+			add_filter( 'wp_update_attachment_metadata',
+				array( $this, 'update_attachment_metadata' ) );
+		}
+
+		add_filter( 'w3tc_preflush_cdn_all',
+			array( $this, 'w3tc_preflush_cdn_all' ), 10, 2 );
 
 		add_filter( 'w3tc_admin_bar_menu',
 			array( $this, 'w3tc_admin_bar_menu' ) );
@@ -89,6 +91,11 @@ class Cdn_Plugin {
 			add_action( 'w3tc_config_ui_save-w3tc_cdn', array(
 					$this, 'change_canonical_header' ), 0, 0 );
 			add_filter( 'w3tc_module_is_running-cdn', array( $this, 'cdn_is_running' ) );
+		}
+
+		if ( !is_admin() || $this->_config->get_boolean( 'cdn.admin.media_library' ) ) {
+			add_filter( 'wp_prepare_attachment_for_js',
+				array( $this, 'wp_prepare_attachment_for_js' ), 0 );
 		}
 
 		/**
@@ -104,35 +111,9 @@ class Cdn_Plugin {
 					'media_row_actions'
 				), 0, 2 );
 		}
-	}
 
-	/**
-	 * run code for FSD CDN
-	 */
-	private function run_fsd() {
-		add_action( 'w3tc_flush_all', array(
-				'\W3TC\Cdn_Fsd_CacheFlush',
-				'w3tc_flush_all'
-			), 3000, 1 );
-		add_action( 'w3tc_flush_post', array(
-				'\W3TC\Cdn_Fsd_CacheFlush',
-				'w3tc_flush_post'
-			), 3000, 1 );
-		add_action( 'w3tc_flushable_posts', '__return_true', 3000 );
-		add_action( 'w3tc_flush_posts', array(
-				'\W3TC\Cdn_Fsd_CacheFlush',
-				'w3tc_flush_all'
-			), 3000 );
-		add_action( 'w3tc_flush_url', array(
-				'\W3TC\Cdn_Fsd_CacheFlush',
-				'w3tc_flush_url'
-			), 3000, 1 );
-		add_filter( 'w3tc_flush_execute_delayed_operations', array(
-				'\W3TC\Cdn_Fsd_CacheFlush',
-				'w3tc_flush_execute_delayed_operations'
-			), 3000 );
-
-		Util_AttachToActions::flush_posts_on_actions();
+		add_filter( 'w3tc_minify_http2_preload_url',
+			array( $this, 'w3tc_minify_http2_preload_url' ), 3000 );
 	}
 
 	/**
@@ -172,6 +153,22 @@ class Cdn_Plugin {
 		$common->upload( $upload, true, $results );
 	}
 
+	function check_inserting_new_attachment( $data, $postarr ) {
+		$this->_attachments_action[ $postarr['file'] ] = empty( $postarr['ID'] ) ? 'insert' : 'update';
+
+		return $data;
+	}
+
+	public function w3tc_preflush_cdn_all( $do_flush, $extras = array() ) {
+		if ( $this->_config->get_boolean( 'cdn.flush_manually' ) ) {
+			if ( !isset( $extras['ui_action'] ) ) {
+				$do_flush = false;
+			}
+		}
+
+		return $do_flush;
+	}
+
 	/**
 	 * Update attachment file
 	 *
@@ -187,7 +184,14 @@ class Cdn_Plugin {
 
 		$results = array();
 
-		$common->upload( $files, true, $results );
+		$cdn_engine = $this->_config->get_string( 'cdn.engine' );
+		if ( Cdn_Util::is_engine_mirror( $cdn_engine ) ) {
+			if ( ! array_key_exists( $attached_file, $this->_attachments_action ) || $this->_attachments_action[ $attached_file ] === 'update' ) {
+				$common->purge( $files, $results );
+			}
+		} else {
+			$common->upload( $files, true, $results );
+		}
 
 		return $attached_file;
 	}
@@ -206,7 +210,12 @@ class Cdn_Plugin {
 
 		$results = array();
 
-		$common->delete( $files, true, $results );
+		$cdn_engine = $this->_config->get_string( 'cdn.engine' );
+		if ( Cdn_Util::is_engine_mirror( $cdn_engine ) ) {
+			$common->purge( $files, $results );
+		} else {
+			$common->delete( $files, true, $results );
+		}
 	}
 
 	/**
@@ -224,7 +233,12 @@ class Cdn_Plugin {
 
 		$results = array();
 
-		$common->upload( $files, true, $results );
+		$cdn_engine = $this->_config->get_string( 'cdn.engine' );
+		if ( Cdn_Util::is_engine_mirror( $cdn_engine ) ) {
+			$common->purge( $files, $results );
+		} else {
+			$common->upload( $files, true, $results );
+		}
 
 		return $metadata;
 	}
@@ -293,11 +307,16 @@ class Cdn_Plugin {
 	 * @return string
 	 */
 	function ob_callback( $buffer ) {
-		if ( $buffer != '' && Util_Content::is_html( $buffer ) ) {
+		if ( $buffer != '' && Util_Content::is_html_xml( $buffer ) ) {
 			if ( $this->can_cdn2( $buffer ) ) {
 				$srcset_helper = new _Cdn_Plugin_ContentFilter();
 				$buffer = $srcset_helper->replace_all_links( $buffer );
-				$this->_replaced_urls = $srcset_helper->get_replaced_urls();
+
+				if ( $this->_debug ) {
+					$replaced_urls = $srcset_helper->get_replaced_urls();
+					$buffer = $this->w3tc_footer_comment_after(
+						$buffer, $replaced_urls );
+				}
 			}
 		}
 
@@ -338,7 +357,7 @@ class Cdn_Plugin {
 	 */
 	function get_files_includes() {
 		$includes_root = Util_Environment::normalize_path( ABSPATH . WPINC );
-		$doc_root = Util_Environment::document_root();
+		$doc_root = Util_Environment::normalize_path( Util_Environment::document_root() );
 		$includes_path = ltrim( str_replace( $doc_root, '', $includes_root ), '/' );
 
 		$files = Cdn_Util::search_files(
@@ -367,7 +386,7 @@ class Cdn_Plugin {
 
 		$themes_root = Util_Environment::normalize_path( $themes_root );
 		$themes_path = ltrim( str_replace(
-				Util_Environment::document_root(), '', $themes_root ), '/' );
+				Util_Environment::normalize_path( Util_Environment::document_root() ), '', $themes_root ), '/' );
 		$files = Cdn_Util::search_files(
 			$themes_root, $themes_path, $this->_config->get_string( 'cdn.theme.files' )
 		);
@@ -391,8 +410,10 @@ class Cdn_Plugin {
 
 			$minify = Dispatcher::component( 'Minify_Plugin' );
 
-			$document_root = Util_Environment::document_root();
-			$minify_root = Util_Environment::cache_blog_dir( 'minify' );
+			$document_root = Util_Environment::normalize_path(
+				Util_Environment::document_root() );
+			$minify_root = Util_Environment::normalize_path(
+				Util_Environment::cache_blog_dir( 'minify' ) );
 			$minify_path = ltrim( str_replace( $document_root, '', $minify_root ), '/' );
 			$urls = $minify->get_urls();
 
@@ -442,14 +463,15 @@ class Cdn_Plugin {
 	 */
 	function get_files_custom() {
 		$files = array();
-		$document_root = Util_Environment::document_root();
+		$document_root = Util_Environment::normalize_path(
+			Util_Environment::document_root() );
 		$custom_files = $this->_config->get_array( 'cdn.custom.files' );
 		$custom_files = array_map( array( '\W3TC\Util_Environment', 'parse_path' ), $custom_files );
-		$site_root = Util_Environment::site_root();
+		$site_root = Util_Environment::normalize_path( Util_Environment::site_root() );
 		$path = Util_Environment::site_url_uri();
 		$site_root_dir = str_replace( $document_root, '', $site_root );
 		if ( strstr( WP_CONTENT_DIR, Util_Environment::site_root() ) === false ) {
-			$site_root = Util_Environment::document_root();
+			$site_root = Util_Environment::normalize_path( Util_Environment::document_root() );
 			$path = '';
 		}
 
@@ -549,15 +571,6 @@ class Cdn_Plugin {
 	 */
 	function can_cdn2( $buffer ) {
 		/**
-		 * Check for database error
-		 */
-		if ( Util_Content::is_database_error( $buffer ) ) {
-			$this->cdn_reject_reason = 'Database Error occurred';
-
-			return false;
-		}
-
-		/**
 		 * Check for DONOTCDN constant
 		 */
 		if ( defined( 'DONOTCDN' ) && DONOTCDN ) {
@@ -613,6 +626,8 @@ class Cdn_Plugin {
 
 		foreach ( $reject_uri as $expr ) {
 			$expr = trim( $expr );
+			$expr = str_replace( '~', '\~', $expr );
+
 			if ( $expr != '' && preg_match( '~' . $expr . '~i', $_SERVER['REQUEST_URI'] ) ) {
 				return false;
 			}
@@ -630,7 +645,7 @@ class Cdn_Plugin {
 	 * @return boolean
 	 */
 	private function _check_logged_in_role_allowed() {
-		global $current_user;
+		$current_user = wp_get_current_user();
 
 		if ( !is_user_logged_in() )
 			return true;
@@ -680,6 +695,82 @@ class Cdn_Plugin {
 		$admin->change_canonical_header();
 	}
 
+	/**
+	 * Adjusts attachment urls to cdn. This is for those who rely on
+	 * wp_prepare_attachment_for_js()
+	 *
+	 * @param 	array   $response	Mixed collection of data about the attachment object
+	 * @return 	array
+	 */
+	public function wp_prepare_attachment_for_js( $response ) {
+		$response['url'] = $this->wp_prepare_attachment_for_js_url( $response['url'] );
+		$response['link'] = $this->wp_prepare_attachment_for_js_url( $response['link'] );
+
+		if ( !empty( $response['sizes'] ) ) {
+			foreach( $response['sizes'] as $size => &$data ) {
+				$data['url'] = $this->wp_prepare_attachment_for_js_url( $data['url'] );
+			}
+		}
+
+		return $response;
+	}
+
+	/**
+	 * An attachment's local url to modify into a cdn url
+	 *
+	 * @param 	string   $url	the local url to modify
+	 * @return 	string
+	 */
+	private function wp_prepare_attachment_for_js_url( $url ) {
+		$url = trim( $url );
+		if ( !empty( $url ) ) {
+			$parsed = parse_url( $url );
+			$uri = ( isset( $parsed['path'] ) ? $parsed['path'] : '/' ) .
+					   ( isset( $parsed['query'] ) ? '?' . $parsed['query'] : '' );
+
+			$wp_upload_dir = wp_upload_dir();
+			$upload_base_url = $wp_upload_dir['baseurl'];
+
+			if ( substr($url, 0, strlen( $upload_base_url ) ) == $upload_base_url ) {
+				$common = Dispatcher::component( 'Cdn_Core' );
+				$new_url = $common->url_to_cdn_url( $url, $uri );
+				if ( !is_null( $new_url ) ) {
+					$url = $new_url;
+				}
+			}
+		}
+
+		return $url;
+	}
+
+	/**
+	 * Mutate http/2 header links
+	 */
+	public function w3tc_minify_http2_preload_url( $data ) {
+		$url = $data['result_link'];
+
+		$url = trim( $url );
+		if ( empty( $url ) ) {
+			return $data;
+		}
+
+		$parsed = parse_url( $url );
+		$uri = ( isset( $parsed['path'] ) ? $parsed['path'] : '/' ) .
+				   ( isset( $parsed['query'] ) ? '?' . $parsed['query'] : '' );
+
+		$common = Dispatcher::component( 'Cdn_Core' );
+		$new_url = $common->url_to_cdn_url( $url, $uri );
+		if ( is_null( $new_url ) ) {
+			return $data;
+		}
+
+		$data['result_link'] = $new_url;
+		// url_to_cdn_url processed by browsercache internally
+		$data['browsercache_processed'] = '*';
+
+		return $data;
+	}
+
 	public function w3tc_admin_bar_menu( $menu_items ) {
 		$cdn_engine = $this->_config->get_string( 'cdn.engine' );
 
@@ -688,7 +779,7 @@ class Cdn_Plugin {
 				'id' => 'w3tc_cdn_flush_all',
 				'parent' => 'w3tc_flush',
 				'title' => __( 'CDN: All', 'w3-total-cache' ),
-				'href' => wp_nonce_url( network_admin_url(
+				'href' => wp_nonce_url( admin_url(
 						'admin.php?page=w3tc_cdn&amp;w3tc_flush_cdn' ),
 					'w3tc' )
 			);
@@ -699,7 +790,7 @@ class Cdn_Plugin {
 				'id' => 'w3tc_cdn_flush',
 				'parent' => 'w3tc_flush',
 				'title' => __( 'CDN: Manual Purge', 'w3-total-cache' ),
-				'href' => wp_nonce_url( network_admin_url( 'admin.php?page=w3tc_cdn&amp;w3tc_cdn_purge' ), 'w3tc' ),
+				'href' => wp_nonce_url( admin_url( 'admin.php?page=w3tc_cdn&amp;w3tc_cdn_purge' ), 'w3tc' ),
 				'meta' => array( 'onclick' => "w3tc_popupadmin_bar(this.href); return false" )
 			);
 		}
@@ -718,28 +809,34 @@ class Cdn_Plugin {
 			( empty( $this->cdn_reject_reason ) ? '' :
 				sprintf( ' (%s)', $this->cdn_reject_reason ) ) );
 
-		if ( $this->_config->get_boolean( 'cdn.debug' ) ) {
-			$strings[] = "CDN debug info:";
-			$strings[] = sprintf( "%s%s", str_pad( 'Engine: ', 20 ),
-				$this->_config->get_string( 'cdn.engine' ) );
-
-			if ( $this->cdn_reject_reason ) {
-				$strings[] = sprintf( "%s%s", str_pad( 'Reject reason: ', 20 ),
-					$this->cdn_reject_reason );
-			}
-
-			if ( count( $this->_replaced_urls ) ) {
-				$strings[] = "Replaced URLs:";
-
-				foreach ( $this->_footer_comment_postfix as $old_url => $new_url ) {
-					$strings[] = sprintf( "%s => %s",
-						Util_Content::escape_comment( $old_url ),
-						Util_Content::escape_comment( $new_url ) );
-				}
-			}
+		if ( $this->_debug ) {
+			$strings[] = '{w3tc_cdn_debug_info}';
 		}
 
 		return $strings;
+	}
+
+
+
+	public function w3tc_footer_comment_after( $buffer, $replaced_urls ) {
+		$strings = array();
+
+		if ( is_array( $replaced_urls ) &&
+				count( $replaced_urls ) ) {
+			$strings[] = "Replaced URLs for CDN:";
+
+			foreach ( $replaced_urls as $old_url => $new_url ) {
+				$strings[] = sprintf( "%s => %s",
+					Util_Content::escape_comment( $old_url ),
+					Util_Content::escape_comment( $new_url ) );
+			}
+
+			$strings[] = '';
+		}
+
+		$buffer = str_replace( '{w3tc_cdn_debug_info}',
+			implode( "\n", $strings ), $buffer );
+		return $buffer;
 	}
 }
 
@@ -748,7 +845,8 @@ class _Cdn_Plugin_ContentFilter {
 	private $_regexps = array();
 	private $_placeholders = array();
 	private $_config;
-	private $_replaced_urls;
+	private $_replaced_urls = array();
+
 	/**
 	 * If background uploading already scheduled
 	 *
@@ -994,27 +1092,58 @@ class _Cdn_Plugin_ContentFilter {
 
 		if ( $this->_config->get_boolean( 'cdn.custom.enable' ) ) {
 			$masks = $this->_config->get_array( 'cdn.custom.files' );
-			$masks = array_map( array( '\W3TC\Cdn_Util', 'replace_folder_placeholders' ), $masks );
+			$masks = array_map( array( '\W3TC\Cdn_Util', 'replace_folder_placeholders_to_uri' ), $masks );
 			$masks = array_map( array( '\W3TC\Util_Environment', 'parse_path' ), $masks );
 
 			if ( count( $masks ) ) {
-				$mask_regexps = array();
+				$custom_regexps_urls = array();
+				$custom_regexps_uris = array();
+				$custom_regexps_docroot_related = array();
 
 				foreach ( $masks as $mask ) {
-					if ( $mask != '' ) {
-						$mask = Util_Environment::normalize_file( $mask );
-						$mask_regexps[] = Cdn_Util::get_regexp_by_mask( $mask );
+					if ( !empty( $mask ) ) {
+						if ( Util_Environment::is_url( $mask ) ) {
+							$url_match = array();
+							if ( preg_match( '~^((https?:)?//([^/]*))(.*)~', $mask, $url_match ) ) {
+								$custom_regexps_urls[] = array(
+									'domain_url' => Util_Environment::get_url_regexp(
+										$url_match[1] ),
+									'uri' => Cdn_Util::get_regexp_by_mask( $url_match[4] )
+								);
+							}
+						} elseif ( substr( $mask, 0, 1 ) == '/' ) {   // uri
+							$custom_regexps_uris[] = Cdn_Util::get_regexp_by_mask( $mask );
+						} else {
+							$file = Util_Environment::normalize_path( $mask );   // \ -> backspaces
+							$file = str_replace( Util_Environment::site_root(), '', $file );
+							$file = ltrim( $file, '/' );
+
+							$custom_regexps_docroot_related[] = Cdn_Util::get_regexp_by_mask( $mask );
+						}
 					}
 				}
 
-				$regexps[] = '~(["\'(=])\s*((' . $domain_url_regexp .
-					')?(' . Util_Environment::preg_quote( $site_path ) .
-					'(' . implode( '|', $mask_regexps ) . ')([^"\'() >]*)))~i';
-				if ( $site_domain_url_regexp )
-					$regexps[] = '~(["\'(=])\s*((' .
-						$site_domain_url_regexp . ')?(' .
-						Util_Environment::preg_quote( $site_path ) . '(' .
-						implode( '|', $mask_regexps ) . ')([^"\'() >]*)))~i';
+				if ( count( $custom_regexps_urls ) > 0 ) {
+					foreach ( $custom_regexps_urls as $regexp ) {
+						$regexps[] = '~(["\'(=])\s*((' . $regexp['domain_url'] .
+						')?((' . $regexp['uri'] . ')([^"\'() >]*)))~i';
+					}
+				}
+				if ( count( $custom_regexps_uris ) > 0 ) {
+					$regexps[] = '~(["\'(=])\s*((' . $domain_url_regexp .
+						')?((' . implode( '|', $custom_regexps_uris ) . ')([^"\'() >]*)))~i';
+				}
+
+				if ( count( $custom_regexps_docroot_related ) > 0 ) {
+					$regexps[] = '~(["\'(=])\s*((' . $domain_url_regexp .
+						')?(' . Util_Environment::preg_quote( $site_path ) .
+						'(' . implode( '|', $custom_regexps_docroot_related ) . ')([^"\'() >]*)))~i';
+					if ( $site_domain_url_regexp )
+						$regexps[] = '~(["\'(=])\s*((' .
+							$site_domain_url_regexp . ')?(' .
+							Util_Environment::preg_quote( $site_path ) . '(' .
+							implode( '|', $custom_regexps_docroot_related ) . ')([^"\'() >]*)))~i';
+				}
 			}
 		}
 
@@ -1037,8 +1166,8 @@ class _Cdn_Plugin_ContentFilter {
 		/**
 		 * Check if URL was already replaced
 		 */
-		if ( isset( $this->replaced_urls[$url] ) ) {
-			return $quote . $this->replaced_urls[$url];
+		if ( isset( $this->_replaced_urls[$url] ) ) {
+			return $quote . $this->_replaced_urls[$url];
 		}
 
 		/**
@@ -1094,16 +1223,9 @@ class _Cdn_Plugin_ContentFilter {
 	 */
 	function _link_replace_callback_ask_cdn( $match, $quote, $url, $path ) {
 		$common = Dispatcher::component( 'Cdn_Core' );
-		$cdn = $common->get_cdn();
-		$remote_path = $common->uri_to_cdn_uri( $path );
-		$new_url = $cdn->format_url( $remote_path );
-		if ( $new_url ) {
-			$is_engine_mirror = Cdn_Util::is_engine_mirror(
-				$this->_config->get_string( 'cdn.engine' ) );
-
-			$new_url = apply_filters( 'w3tc_cdn_url', $new_url, $url,
-				$is_engine_mirror );
-			$this->replaced_urls[$url] = $new_url;
+		$new_url = $common->url_to_cdn_url( $url, $path );
+		if ( !is_null( $new_url ) ) {
+			$this->_replaced_urls[$url] = $new_url;
 			return $quote . $new_url;
 		}
 
@@ -1146,17 +1268,6 @@ class _Cdn_Plugin_ContentFilter {
 	}
 
 	function get_replaced_urls() {
-		$strings = array();
-		if ( count( $this->_replaced_urls ) ) {
-			$strings[] = "Replaced URLs:";
-
-			foreach ( $this->_replaced_urls as $old_url => $new_url ) {
-				$strings[] = sprintf( "%s => %s",
-					Util_Content::escape_comment( $old_url ),
-					Util_Content::escape_comment( $new_url ) );
-			}
-		}
-		return $strings;
+		return $this->_replaced_urls;
 	}
-
 }
